measurement conclusion 7mar2023
measurements were taken during this cool down of the small glass dewar. more code should be written to assist in better retention of electrons. however, this ends the cooldown and, thus, the necessity for the branch. Once the system warms up and the temperature script is stopped, then the branch will be merged.